automation_framework/requirements.txt
selenium==4.18.1
pytest==8.1.1
pytest-html==4.1.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest-xdist==3.5.0
webdriver-manager==4.0.1
